fix: corrigindo modulação do dangerfile.js
